+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+## [1.2.0] -- 2020-08-*
+
+This release includes some extension functions to make code more readable.
+
+### Added
+* Added `Either<R>.onRight` -- extension function to handle `Right` case
+* Added `Either<R>.onSuccess` -- extension function to handle `Right` case
+* Added `Either<L>.onLeft` -- extension function to handle `Left` case
+* Added `Either<L>.onFailure` -- extension function to handle `Left` case
+
 ## [1.1.0] -- 2020-04-29
 
 ### Added
@@ -29,7 +39,8 @@
 ### Added
 * Initial release
 
-[Unreleased]: https://github.com/sphrak/Either/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/sphrak/Either/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/sphrak/Either/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/sphrak/Either/compare/1.0.2...1.1.0
 [1.0.2]: https://github.com/sphrak/Either/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/sphrak/Either/compare/1.0.0...1.0.1